requirements.txt
Django==4.2.16
Pillow==10.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==22.0.0

# Reportes y exportación
reportlab==4.0.9
openpyxl==3.1.2
xlsxwriter==3.2.0

# APIs y serialización
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Utilidades adicionales
django-filter==23.5
django-tables2==2.7.0

# Testing dependencies
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
coverage==7.4.4
django-test-plus==2.3.0

